Renamed the CreateSecurityGroupTemplate to CreateSecurityGroupRequest class and the SecurityRuleRequest to Rule class
